pyproject.toml
[project]
name = "chorus"
version = "0.1.0"
description = "A script to manage multiple copies of multiple repos being edited by coding agents in parallel"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click",
    "pyyaml",
    "rich",
    "questionary",
    "libtmux>=0.46.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[project.scripts]
chorus = "chorus.cli:main"
